UART: Add com_rst and improve interface thread
The proper way to reset the communication is using com_rst. It resets
all buffers and local state of the communication infrastructure,
except the receive line handling that will not be reset as it handles
the reset registers.

Improve the interface thread by reading and writing with a local
buffer. For writes we just plainly foward data, for reads we need to
parse for credits which goes into an extra function.