UART: Add extra buffer for flow control
It is necessary to track the actual transfers out of the ingress
FIFO. As we use this for clock domain crossing, it would have been
possible to cross the read counter of the FIFO again to the IO clock
and track this one for the transfer counter.

But for the moment it seems much more flexible to use an extra 4kB
buffer just for the flow control, which is one extra BRAM. Taking we
are not using many resources, this should not be a problem.