TCP: Use singleton for DPI
To be able to check the status of the Glip DPI interface, we need to
call the GlipTcp class instance. The most convenient way is by making
the GlipTcp a singleton and thereby limit it to one instance.

This can be changed later by using svScope and maintaining a list of
all instances. But as we consider the usage of multiple Glip
interfaces in one design a rare use case, we stick to the singleton
for now.