UART: Move to "new" style of custom baud rate
The serial interface is deprecated and a termio2 data structure should
instead be used. It has two fields for the input and output baud rate
and BOTHER (B-OTHER) needs to be set in the c_flag.

We always use a common baud rate and spare the known baud rates
entirely.