UART: Change module ordering in ingress data path
The main motiviation of this change is to improve the timing of the
ingress path. We move the ingress buffer directly to the user logic as
last element in the chain, the place where the upscaler has been sitting
before. This ensures that the fifo_in_* signals are a clean FIFO
interface with minimal combinatorial logic, thus users have more freedom
to connect their own logic on this interface.

To make such changes easier in the future, the code was refactored in
the following way:
- In module instantiations only wires named after the ports are
  connected.
- The connection logic is put in separate assign statements.
- The conditional logic of the upscaler/downscaler depending on the
  input/output width ratio has been moved into the upscaler/downscaler
  itself. This makes the toplevel easier to read, and paves the way for
  future extensions to up/down-scaling factors other than 1:2/2:1.

With these changes, the timing errors when using GLIP inside Open SoC
Debug on the VCU108 board are gone.

Fixes #32