Replace FIFO implementation with OpTiMSoC one
The old FIFO implementation is put together from various sources for
GLIP. Use the new implementation from OpTiMSoC to share the testing and
the optimizations later on.

Currently the files are copied from the OpTiMSoC source tree, if fusesoc
evolves further we might be able reduce this copying in the future.