Added 16 and 32 bit version of FX3 toplevel.
The FX3 interface and the interface to the FPGA logic now have the same width.
Two different firmware versions are available for the FX3, one with a 16 bit
wide interface to the FPGA and one with a 32 bit wide interface.
The FIFO width can be read out by the glip backend.