cypressfx3: make fx3_pclk an output port
Currently the toplevel needs to create the fx3_pclk and pass it then to
the output port of the design, as well as to the glip toplevel. This is
confusing, as it doesn't follow what (for example) the UART backend
does, or how the fx3_* signals are used in all other cases.

This commit introduces a clk_io (as in UART), which is then passed
through as fx3_pclk, and used internally on the FX3 side.