Draft FX3 firmware interface
It not make sense to load our firmware to the ZTEX boards, because we
would loose the ability to load the FPGA bitstream. Also we don't want
to include this in our firmware. Finally, we want to support the
default ZTEX firmware.

Hence, this drafts one possibility to allow for different firmware
interfaces. The differences between interfaces are:

 * How GPIO pins are accessed

 * Which endpoints are used for the FIFOs

If we anticipate only ZTEX and our firmware, I would suggest we
explore adopting the ZTEX interface. Else this is maybe a direction we
want to go.