Fixed fx3 flush bug.
The bug appeared when the data that was written to the fx3 was flushed and new data
to be sent out was available before the 'fx3_in_full' flag was set by the fx3.

The partial empty/full flags were adjusted in the firmware. This allows for a simpler
FSM on the FPGA which can handle 16-bit and 32-bit interfaces the same way now.

Furthermore, the io clock must now always be 100MHz to avoid timing problems.
Timing constraints for the I/O pins have been added to the hpc constraint file.